Fix the unit tests.
(Not like anyone ever runs them...)

`make check` will now backup config.xml and replace it with config.xml.check,
so that when running lb.exe we'll get a consistent config.xml file (as changes
to config.xml cause changes to the test output, which defeats the purpose of
testing the output...).

Once `make check` finishes, we replace config.xml with the original contents.

svn path=/trunk/lb/; revision=149186